Major refactoring for Python3
